requirements.txt
# Telegram Bot Framework
aiogram==3.20.0.post0

# Async HTTP Client
httpx==0.28.1

# Async File Operations
aiofiles==24.1.0

# Database
SQLAlchemy==2.0.40
asyncpg==0.30.0

# Data Processing
pandas==2.2.3
numpy>=1.24.0
openpyxl==3.1.5

# Environment Variables
python-dotenv==1.1.0

# Data Validation
pydantic==2.11.3

# Payment Integration (Модуль Банк)
aiohttp==3.9.5

# Security - Token Encryption
cryptography==44.0.0

# Aiogram Dependencies
magic-filter==1.0.12
propcache==0.3.1
